Changes for Issue 81
Moves FileLogManagerOptions to ServerOptions.
