+++ src/main/java/com/tesco/mewbase/log/impl/file/FileLogManager.java
@@ -3,6 +3,7 @@ package com.tesco.mewbase.log.impl.file;
 import com.tesco.mewbase.client.MewException;
 import com.tesco.mewbase.log.Log;
 import com.tesco.mewbase.log.LogManager;
+import com.tesco.mewbase.server.ServerOptions;
 import io.vertx.core.Vertx;
 
 import java.io.File;
@@ -18,10 +19,10 @@ public class FileLogManager implements LogManager {
     private final Vertx vertx;
     private final FileAccess faf;
     private final File logDir;
-    private final FileLogManagerOptions options;
     private final Map<String, FileLog> logs = new ConcurrentHashMap<>();
+    private final ServerOptions options;
 
-    public FileLogManager(Vertx vertx, FileLogManagerOptions options, FileAccess faf) {
+    public FileLogManager(Vertx vertx, ServerOptions options, FileAccess faf) {
         this.vertx = vertx;
         this.logDir = new File(options.getLogDir());
         if (!logDir.exists()) {
