+++ src/test/java/com/tesco/mewbase/SSLPubSubTest.java
@@ -5,7 +5,6 @@ import com.tesco.mewbase.client.ClientDelivery;
 import com.tesco.mewbase.client.ClientOptions;
 import com.tesco.mewbase.client.Producer;
 import com.tesco.mewbase.common.SubDescriptor;
-import com.tesco.mewbase.log.impl.file.FileLogManagerOptions;
 import com.tesco.mewbase.server.ServerOptions;
 import io.vertx.core.net.NetClientOptions;
 import io.vertx.core.net.PemKeyCertOptions;
@@ -27,17 +26,13 @@ public class SSLPubSubTest extends ServerTestBase {
 
     @Override
     protected ServerOptions createServerOptions(File logDir) {
-        FileLogManagerOptions fileLogManagerOptions = new FileLogManagerOptions().setLogDir(logDir.getPath());
-
         ServerOptions serverOptions = new ServerOptions().setChannels(new String[]{TEST_CHANNEL_1, TEST_CHANNEL_2})
-                .setFileLogManagerOptions(fileLogManagerOptions);
-
+        		.setLogDir(logDir.getPath());  
         serverOptions.getNetServerOptions().setSsl(true).setPemKeyCertOptions(
                 new PemKeyCertOptions()
                         .setKeyPath(KEY_PATH)
                         .setCertPath(CERT_PATH)
         );
-
         return serverOptions;
     }
